Improve sanctuary JS pipe
On suggestion of sanctuary js designer. This uses
S.prop instead of S.value, which is the more appropriate
choice since the data here is a "known" object. You use
S.value with StrMap, which has homoagenous values. And
S.get with unknown objects. But this is a known object
b/c coming from the API query.