Add resources page (#4)
* Add designs with resource page mock

* Upgrade deps

* Add unfinished filter component with working query

* Add rough version of filters and tag components

* Add progress on Resource page

* Add updated style sketches

* Add counts to filter tags

This commit also builds the filter set data from
the query into the resources posts.

* Improve sanctuary JS pipe

On suggestion of sanctuary js designer. This uses
S.prop instead of S.value, which is the more appropriate
choice since the data here is a "known" object. You use
S.value with StrMap, which has homoagenous values. And
S.get with unknown objects. But this is a known object
b/c coming from the API query.

* Clean up

* Add comment

* Simplify dataToResources pipe

Can just use `S.prop` because this is an API
object whose structure is "known". If these props
aren't present, we want program to crash.

* Clean up unused

* Add filtering mechanism

* Add mechanism to mark selected filter tag

* Add link to contentTeaser module

* Add --text-font size variable

* Update styles for contentTeaser module

* Add missing variables import

* Clean up

* Update contentTeaser styles

* Fix import